Refactor code to use command end marker
The code has been refactored to support a command end marker. An empty
serverdata_value packet with the same ID as the command packet that
comes before it. The reply to that serverdata_value packet serves as an
end marker for the command replies. This allows to properly act upon
commands that send huge amounts of data such as "cvarlist".