Add support for disabling multi-packet responses
Some servers (e.g. Factorio) don't implement the trick of adding a
0-length SERVERDATA_RESPONSE_VALUE message at the end of messages
because they never try to send multi-packet responses. This commit adds
a new flag that effectively disables any expectation of multi-packet
responses, and in particular makes it possible to use it with the
headless Factorio server (as of versions <= 0.15).